* I tested these changes real quick in game (CMT's b30) and in Guerilla * VS2012.2's intellisense is broken, ignore it. http://connect.microsoft.com/VisualStudio/feedback/details/745237/invalid-byte-order-in-intellisense
* blamlib: All tag_group related functions now sit in the 'blam' namespace
* YeloLib: Added range-based for support to TagBlock<T> and DatumArray<T>
* Halo1: Reimplemented all simple for(;;) based loops on static arrays using range based fors
* Halo1: Started using std::array in place of static C arrays (functionally they're the same)
* Halo1_CE: Changed all instances where DataArray::Iterator is used and implemented with range-based fors
* Halo1_CE: Reimplemented all data_array getters as T&
* Halo1_CheApe: Started work on RE'ing the loading code used for tag_groups (triva: tag_group_loading.cpp didn't appear until H2)

I may be missing a few other minor changes. Sue me!

+++ OpenSauce/Halo1/Halo1_CE/TagGroups/CacheFiles.MapList.inl
@@ -10,7 +10,7 @@ cstring* MapListIgnoredMapNames()				PTR_IMP_GET2(map_list_ignored_map_names);
 
 static void MapListInitialize()
 {
-	static const uint32 INITIALIZE = GET_FUNC_PTR(MULTIPLAYER_MAP_LIST_INITIALIZE);
+	static const uintptr_t INITIALIZE = GET_FUNC_PTR(MULTIPLAYER_MAP_LIST_INITIALIZE);
 
 	__asm	call	INITIALIZE
 	MapListInitializeYelo();
@@ -18,7 +18,7 @@ static void MapListInitialize()
 
 API_FUNC_NAKED void MapListReIntialize()
 {
-	static const uint32 DISPOSE = GET_FUNC_PTR(MULTIPLAYER_MAP_LIST_DISPOSE);
+	static const uintptr_t DISPOSE = GET_FUNC_PTR(MULTIPLAYER_MAP_LIST_DISPOSE);
 
 	__asm {
 		call	DISPOSE
